+++ packages/devtools_app/lib/src/framework/home_screen.dart
@@ -62,8 +62,9 @@ class _HomeScreenBodyState extends State<HomeScreenBody> with AutoDisposeMixin {
 
   @override
   Widget build(BuildContext context) {
-    final connected = serviceConnection.serviceManager.hasConnection &&
-        serviceConnection.serviceManager.connectedAppInitialized;
+    final connected =
+        serviceConnection.serviceManager.connectedState.value.connected &&
+            serviceConnection.serviceManager.connectedAppInitialized;
     return Scrollbar(
       child: ListView(
         children: [
@@ -102,6 +103,9 @@ class ConnectionSection extends StatelessWidget {
             gaScreen: gac.home,
             minScreenWidthForTextBeforeScaling:
                 _primaryMinScreenWidthForTextBeforeScaling,
+            routerDelegate: DevToolsRouterDelegate.of(context),
+            onPressed: () =>
+                Navigator.of(context, rootNavigator: true).pop('dialog'),
           ),
         ],
         child: const ConnectedAppSummary(narrowView: false),