+++ packages/devtools_app/lib/src/screens/inspector/inspector_screen.dart
@@ -129,7 +129,7 @@ class InspectorScreenBodyState extends State<InspectorScreenBody>
       }
     });
     addAutoDisposeListener(preferences.inspector.pubRootDirectories, () {
-      if (serviceConnection.serviceManager.hasConnection &&
+      if (serviceConnection.serviceManager.connectedState.value.connected &&
           controller.firstInspectorTreeLoadCompleted) {
         _refreshInspector();
       }